requirements.txt
pyautogui
keyboard
pygetwindow
opencv-python
pillow
tk
